pyproject.toml
[tool.poetry]
name = "ape-scratch"
version = "0.1.0"
authors = ["ArWer <arwer13@gmail.com>"]
description = ""

[tool.poetry.dependencies]
python = ">=3.10.4,<3.11"
eth-ape = "^0.6.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
